# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CyberOpus API
  description: |
    CyberOpus Online Bookstore REST API.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained from `/api/auth/login` or `/api/auth/register`.
    Include it as: `Authorization: Bearer <token>`

    ## Test Payment
    Use `cardHolderName: "FAIL_TEST"` to simulate a payment failure.
    Normal names simulate success.
  version: "1.0"
  contact:
    name: CyberOpus
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        role: { type: string, enum: [CUSTOMER, ADMIN] }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/UserResponse' }
    RegisterRequest:
      type: object
      required: [firstName, lastName, email, password]
      properties:
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, description: "Must contain uppercase, lowercase, digit, and special character" }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    CategoryResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string }
        imageUrl: { type: string }
    BrandResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string }
        logoUrl: { type: string }
    BookResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        author: { type: string }
        description: { type: string }
        price: { type: number, format: double }
        stockQuantity: { type: integer }
        isbn: { type: string }
        imageUrl: { type: string }
        publicationYear: { type: integer }
        active: { type: boolean }
        category: { $ref: '#/components/schemas/CategoryResponse' }
        brand: { $ref: '#/components/schemas/BrandResponse' }
        salesCount: { type: integer, format: int64 }
        deliveryDate: { type: string, format: date }
        inStock: { type: boolean }
    PageResponseBook:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/BookResponse' }
        page: { type: integer }
        size: { type: integer }
        totalPages: { type: integer }
        totalElements: { type: integer, format: int64 }
...
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new customer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Registered successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '409':
          description: Email already in use
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '401':
          description: Not authenticated
  /api/books:
    get:
      tags: [Books]
      summary: Get books with filtering and pagination
      security: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 12 }
        - in: query
          name: sort
          schema: { type: string, enum: [price_asc, price_desc, title, newest, bestseller] }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category
          schema: { type: integer }
        - in: query
          name: brand
          schema: { type: integer }
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: inStock
          schema: { type: boolean }
      responses:
        '200':
          description: Paginated book list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageResponseBook' }
    post:
      tags: [Books, Admin]
      summary: Create a new book (ADMIN)
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookResponse' }
        '403':
          description: Admin only
  /api/books/{id}:
    get:
      tags: [Books]
      summary: Get book by ID
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookResponse' }
        '404':
          description: Book not found
    put:
      tags: [Books, Admin]
      summary: Update book (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Updated book
    delete:
      tags: [Books, Admin]
      summary: Delete book (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Deleted
  /api/books/search:
    get:
      tags: [Books]
      summary: Search books by query
      security: []
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Matching books
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageResponseBook' }
  /api/books/{id}/related:
    get:
      tags: [Books]
      summary: Get related books
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Related books list
...
  /api/brands/{id}:
    get:
      tags: [Brands]
      summary: Get brand by ID
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Brand
    put:
      tags: [Brands, Admin]
      summary: Update brand (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Updated brand
    delete:
      tags: [Brands, Admin]
      summary: Delete brand (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Deleted
  /api/brands/{id}/books:
    get:
      tags: [Brands]
      summary: Get books by brand
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Books by brand
...
  /api/cart/items:
    post:
      tags: [Cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId, quantity]
              properties:
                bookId: { type: integer, format: int64 }
                quantity: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Updated cart
  /api/cart/items/{itemId}:
    put:
      tags: [Cart]
      summary: Update cart item quantity
      parameters:
        - in: path
          name: itemId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Updated cart
    delete:
      tags: [Cart]
      summary: Remove item from cart
      parameters:
        - in: path
          name: itemId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Updated cart
...
  /api/addresses/{id}:
    put:
      tags: [Addresses]
      summary: Update address
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Updated address
    delete:
      tags: [Addresses]
      summary: Delete address
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Deleted
  /api/addresses/{id}/default:
    patch:
      tags: [Addresses]
      summary: Set address as default
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Address set as default
...